goldens and the C++ battery for the eight-byte slot
Only the POINTERED unit moves, which is what §6.3 now claims: tables-pointers'
build version and its three generated headers. The value-only corpus, the
block corpus and every wire golden are byte-identical — the wire is untouched
by a region encoding, and a table with no pointer has no slot.

test/tables/main.cpp pinned the slot at four bytes and went red, which is the
gate doing its job. Re-pinned at eight, and strengthened while it is open: the
alignment, null in both encodings, a NEGATIVE delta as an ordinary value (the
back-reference the old comment denied), and the pointer field's offset against
the alignment a cook writes at. The C++ battery passes — Lock, Save and Load
over the pointered corpus included.

+++ testdata/golden/tables/pointers/PartsTable.h
@@ -75,7 +75,7 @@ struct TableFieldInfo
     uint16_t id;            // table-wire field id (name hash; the was alias's hash after a rename)
     uint8_t kind;           // table-wire kind; for arrays/strings/bytes, the ELEMENT kind
     bool is_array;          // fixed or counted array (bytes included)
-    bool is_pointer;        // a *T pointer field: storage is a 4-byte TableRef; the target is a table
+    bool is_pointer;        // a *T pointer field: storage is an 8-byte TableRef; the target is a table
     bool counted;           // a _count/_length int32 companion exists (counted arrays, strings, bytes)
     bool optional;          // a ?T field: a _present bool companion decides whether it rides
     int32_t array_bound;    // array capacity / string max length; 0 for plain scalars
@@ -340,12 +340,21 @@ static const int32_t kTableMaxDepth = 128;
 //                   so a deref is one add, needs no base pointer, and a whole
 //                   region relocates by memcpy with zero fix-up
 //
-// 0 is null in both. Region deltas are always POSITIVE: a region is packed by
-// a depth-first walk, so a child always sits after the slot that names it —
-// which is what makes a packed region cycle-free by construction.
+// 0 is null in both, and a slot can never name the node that contains it, so
+// zero names nothing real in either form.
+//
+// A REGION DELTA HAS NO REQUIRED SIGN (§6.3). A region is packed depth-first,
+// so a node's FIRST reference points forward; every LATER reference to that
+// same node points BACK at the one body it already has, which is exactly what
+// makes one node one node in a region. Sharing and a back-reference are the
+// same fact, and nothing validates a reference by its sign.
+//
+// IT IS EIGHT BYTES, SIGNED, so ONE REGION REACHES EVERYTHING (§6.3, §7): a
+// four-byte slot bounded a region at 2 GiB, and the scale a cook exists for is
+// *"100mbs or many gigabytes of data in Assets.bin"*.
 struct TableRef
 {
-    uint32_t value = 0;
+    int64_t value = 0;
     bool null() const { return value == 0; }
 };
 
@@ -758,7 +767,7 @@ inline bool ColourLoad( Colour & value, const uint8_t * buffer, int64_t bytes, T
 // memcpy'd, mmap'd, shared across processes, and walked through
 // descriptor offsets. A failure here means a pointer, virtual or
 // non-trivial member crept into generated storage.
-// A pointer FIELD is a TableRef — four bytes and no address — so the
+// A pointer FIELD is a TableRef — eight bytes and no address — so the
 // property holds in BOTH forms: a fixed-size table is one relocatable
 // struct, and a packed region is one relocatable block whose references
 // are self-relative and therefore survive a plain memcpy.
